refactor(searchMonitor): Add pagination to searchMonitors
The searchMonitors function did not properly handle pagination. With the
default page size being 30 serverless stacks with > 30 monitors always
encounter a 400 when updating monitors because some of the monitors will
not be returned in the search and will not be listed in existingMonitors.
This pr adds pagination to the searchMonitors function to fix this bug.