Search monitor patch (#433)
* refactor(searchMonitor): Add pagination to searchMonitors

The searchMonitors function did not properly handle pagination. With the
default page size being 30 serverless stacks with > 30 monitors always
encounter a 400 when updating monitors because some of the monitors will
not be returned in the search and will not be listed in existingMonitors.
This pr adds pagination to the searchMonitors function to fix this bug.

* refactor(searchMonitor): Update searchMonitor tests

Update the searchMonitor tests to ensure pagination is working and will
make multiple calls to the api if needed.

* chore(lint): Fix linting errors

Ran npm run format to fix linting errors.